First round of comming to a generic ui
