+++ css/main.css
@@ -4,6 +4,210 @@
   --OR-margin-50: 50px;
 }
 
+/* Card Components */
+
+/* Card Grid Layouts */
+.cardGrid {
+  display: grid;
+  gap: 1.5rem;
+  grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+}
+
+.chartGrid {
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+  gap: 20px;
+  padding: 20px;
+}
+
+/* Base Card Styles */
+.card {
+  background: var(--color-main-background);
+  border-radius: 8px;
+  padding: 20px;
+  box-shadow: 0 2px 8px var(--color-box-shadow);
+  min-height: 200px;
+  transition: transform 0.2s ease-in-out;
+  border: 1px solid var(--color-border);
+}
+
+.card:hover {
+  transform: scale(1.01);
+  box-shadow: 0 4px 12px var(--color-box-shadow);
+}
+
+.cardHeader {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 8px;
+  margin-bottom: 12px;
+  padding-bottom: 8px;
+  border-bottom: 1px solid var(--color-border);
+}
+
+.cardHeader h2,
+.cardHeader h3 {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  margin: 0;
+  color: var(--color-main-text);
+}
+
+.cardHeader h2 {
+  font-size: 1.2em;
+}
+
+.cardHeader h3 {
+  font-size: 1.1em;
+}
+
+/* Chart Card Specific */
+.chartCard {
+  background: var(--color-main-background);
+  border-radius: 8px;
+  padding: 20px;
+  box-shadow: 0 2px 8px var(--color-box-shadow);
+  border: 1px solid var(--color-border);
+}
+
+.chartCard h3 {
+  margin: 0 0 20px 0;
+  font-size: 1.2em;
+  color: var(--color-main-text);
+}
+
+/* Nested Card Components */
+.nestedCardContainer {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  margin-top: 20px;
+  padding-top: 16px;
+  border-top: 1px solid var(--color-border);
+}
+
+.nestedCard {
+  border: 1px solid var(--color-border);
+  border-radius: var(--border-radius);
+  margin-bottom: 8px;
+  background-color: var(--color-main-background);
+}
+
+.nestedCard:last-child {
+  margin-bottom: 0;
+}
+
+.nestedCardHeader {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 8px 12px;
+  cursor: pointer;
+  transition: background-color 0.2s ease;
+}
+
+.nestedCardHeader:hover {
+  background-color: var(--color-background-hover);
+}
+
+.nestedCardTitle {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  font-size: 0.9em;
+}
+
+.nestedCardContent {
+  border-top: 1px solid var(--color-border);
+  background-color: var(--color-background-hover);
+  padding: 12px;
+}
+
+/* Stat Components */
+.statGrid {
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+  gap: 12px;
+  margin-bottom: 16px;
+}
+
+.statItem {
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+}
+
+.statLabel {
+  color: var(--color-text-maxcontrast);
+  font-size: 0.9em;
+}
+
+.statValue {
+  font-size: 1.1em;
+  font-weight: 600;
+}
+
+/* Toggle Buttons */
+.toggleButton {
+  background: none;
+  border: 1px solid var(--color-border);
+  border-radius: var(--border-radius);
+  padding: 4px 8px;
+  cursor: pointer;
+  color: var(--color-text-maxcontrast);
+  font-size: 0.9em;
+  transition: all 0.2s ease;
+}
+
+.toggleButton:hover {
+  color: var(--color-main-text);
+  background-color: var(--color-background-hover);
+}
+
+/* Content and Layout Containers */
+.dashboardContent {
+  margin-inline: auto;
+  max-width: 1200px;
+  padding-block: 20px;
+  padding-inline: 20px;
+}
+
+.loadingContainer {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  color: var(--color-text-maxcontrast);
+  justify-content: center;
+  padding-block: 40px;
+}
+
+/* Header Actions */
+.headerActionsContainer {
+  display: flex;
+  align-items: end;
+  gap: 1em;
+  padding-block-start: 0.5em;
+  padding-inline-end: 0.9em;
+}
+
+.viewModeSwitchContainer {
+  display: flex;
+  align-items: center;
+}
+
+.viewModeSwitchContainer span {
+  max-height: 34px;
+}
+
+/* Responsive Design */
+@media screen and (max-width: 1024px) {
+  .chartGrid {
+    grid-template-columns: 1fr;
+  }
+}
+
 /* Pages */
 
 .pageHeader {
