+++ src/views/schema/SchemasIndex.vue
@@ -5,293 +5,317 @@ import formatBytes from '../../services/formatBytes.js'
 
 <template>
 	<NcAppContent>
-		<span class="pageHeaderContainer">
-			<h2 class="pageHeader">
-				Schemas
-			</h2>
+		<div class="viewContainer">
+			<!-- Header -->
+			<div class="viewHeader">
+				<h1 class="viewHeaderTitleIndented">
+					{{ t('openregister', 'Schemas') }}
+				</h1>
+				<p>{{ t('openregister', 'Manage your data schemas and their properties') }}</p>
+			</div>
 
-			<div class="headerActionsContainer">
-				<div class="viewModeSwitchContainer">
-					<NcCheckboxRadioSwitch
-						v-model="schemaStore.viewMode"
-						v-tooltip="'See schemas as cards'"
-						:button-variant="true"
-						value="cards"
-						name="view_mode_radio"
-						type="radio"
-						button-variant-grouped="horizontal">
-						Cards
-					</NcCheckboxRadioSwitch>
-					<NcCheckboxRadioSwitch
-						v-model="schemaStore.viewMode"
-						v-tooltip="'See schemas as a table'"
-						:button-variant="true"
-						value="table"
-						name="view_mode_radio"
-						type="radio"
-						button-variant-grouped="horizontal">
-						Table
-					</NcCheckboxRadioSwitch>
+			<!-- Actions Bar -->
+			<div class="viewActionsBar">
+				<div class="viewInfo">
+					<span class="viewTotalCount">
+						{{ t('openregister', 'Showing {count} schemas', { count: schemaStore.schemaList.length }) }}
+					</span>
 				</div>
+				<div class="viewActions">
+					<div class="viewModeSwitchContainer">
+						<NcCheckboxRadioSwitch
+							v-model="schemaStore.viewMode"
+							v-tooltip="'See schemas as cards'"
+							:button-variant="true"
+							value="cards"
+							name="view_mode_radio"
+							type="radio"
+							button-variant-grouped="horizontal">
+							Cards
+						</NcCheckboxRadioSwitch>
+						<NcCheckboxRadioSwitch
+							v-model="schemaStore.viewMode"
+							v-tooltip="'See schemas as a table'"
+							:button-variant="true"
+							value="table"
+							name="view_mode_radio"
+							type="radio"
+							button-variant-grouped="horizontal">
+							Table
+						</NcCheckboxRadioSwitch>
+					</div>
 
-				<NcActions
-					:force-name="true"
-					:inline="1"
-					:primary="true"
-					menu-name="Schema actions">
-					<NcActionButton close-after-click @click="schemaStore.setSchemaItem(null); navigationStore.setModal('editSchema')">
-						<template #icon>
-							<PlusCircleOutline :size="20" />
-						</template>
-						Add Schema
-					</NcActionButton>
-					<NcActionButton close-after-click @click="schemaStore.refreshSchemaList()">
-						<template #icon>
-							<Refresh :size="20" />
-						</template>
-						Refresh
-					</NcActionButton>
-				</NcActions>
-			</div>
-		</span>
-
-		<div class="dashboardContent">
-			<div v-if="!schemaStore.schemaList.length" class="empty">
-				<NcEmptyContent title="No schemas found" icon="icon-folder" />
-			</div>
-			<div v-else class="registers">
-				<template v-if="schemaStore.viewMode === 'cards'">
-					<div v-for="schema in schemaStore.schemaList" :key="schema.id" class="registerCard">
-						<div class="registerHeader">
-							<h2>
-								<FileTreeOutline :size="20" />
-								{{ schema.title }}
-							</h2>
-							<NcActions :primary="true" menu-name="Actions">
-								<template #icon>
-									<DotsHorizontal :size="20" />
-								</template>
-								<NcActionButton close-after-click @click="schemaStore.setSchemaItem(schema); navigationStore.setModal('editSchema')">
-									<template #icon>
-										<Pencil :size="20" />
-									</template>
-									Edit
-								</NcActionButton>
-								<NcActionButton close-after-click @click="schemaStore.setSchemaPropertyKey(null); schemaStore.setSchemaItem(schema); navigationStore.setModal('editSchemaProperty')">
-									<template #icon>
-										<PlusCircleOutline :size="20" />
-									</template>
-									Add Property
-								</NcActionButton>
-								<NcActionButton close-after-click @click="schemaStore.downloadSchema(schema)">
-									<template #icon>
-										<Download :size="20" />
-									</template>
-									Download
-								</NcActionButton>
-								<NcActionButton v-tooltip="schema.stats?.objects?.total > 0 ? 'Cannot delete: objects are still attached' : ''"
-									close-after-click
-									:disabled="schema.stats?.objects?.total > 0"
-									@click="schemaStore.setSchemaItem(schema); navigationStore.setDialog('deleteSchema')">
-									<template #icon>
-										<TrashCanOutline :size="20" />
-									</template>
-									Delete
-								</NcActionButton>
-								<NcActionButton close-after-click @click="schemaStore.setSchemaItem(schema); navigationStore.setSelected('schemaDetails')">
-									<template #icon>
-										<InformationOutline :size="20" />
-									</template>
-									View Details
-								</NcActionButton>
-							</NcActions>
-						</div>
-						<!-- Toggle between stats and properties -->
-						<div v-if="!schema.showProperties">
-							<table class="statisticsTable schemaStats">
-								<thead>
-									<tr>
-										<th>{{ t('openregister', 'Type') }}</th>
-										<th>{{ t('openregister', 'Total') }}</th>
-										<th>{{ t('openregister', 'Size') }}</th>
-									</tr>
-								</thead>
-								<tbody>
-									<tr>
-										<td>{{ t('openregister', 'Registers') }}</td>
-										<td>{{ schema.stats?.registers ?? 0 }}</td>
-										<td>-</td>
-									</tr>
-									<tr>
-										<td>{{ t('openregister', 'Properties') }}</td>
-										<td>{{ Object.keys(schema.properties).length }}</td>
-										<td>-</td>
-									</tr>
-									<tr>
-										<td>{{ t('openregister', 'Objects') }}</td>
-										<td>{{ schema.stats?.objects?.total || 0 }}</td>
-										<td>{{ formatBytes(schema.stats?.objects?.size || 0) }}</td>
-									</tr>
-									<tr class="subRow">
-										<td class="indented">
-											{{ t('openregister', 'Invalid') }}
-										</td>
-										<td>{{ schema.stats?.objects?.invalid || 0 }}</td>
-										<td>-</td>
-									</tr>
-									<tr class="subRow">
-										<td class="indented">
-											{{ t('openregister', 'Deleted') }}
-										</td>
-										<td>{{ schema.stats?.objects?.deleted || 0 }}</td>
-										<td>-</td>
-									</tr>
-									<tr class="subRow">
-										<td class="indented">
-											{{ t('openregister', 'Locked') }}
-										</td>
-										<td>{{ schema.stats?.objects?.locked || 0 }}</td>
-										<td>-</td>
-									</tr>
-									<tr class="subRow">
-										<td class="indented">
-											{{ t('openregister', 'Published') }}
-										</td>
-										<td>{{ schema.stats?.objects?.published || 0 }}</td>
-										<td>-</td>
-									</tr>
-									<tr>
-										<td>{{ t('openregister', 'Logs') }}</td>
-										<td>{{ schema.stats?.logs?.total || 0 }}</td>
-										<td>{{ formatBytes(schema.stats?.logs?.size || 0) }}</td>
-									</tr>
-									<tr>
-										<td>{{ t('openregister', 'Files') }}</td>
-										<td>{{ schema.stats?.files?.total || 0 }}</td>
-										<td>{{ formatBytes(schema.stats?.files?.size || 0) }}</td>
-									</tr>
-								</tbody>
-							</table>
-						</div>
-						<div v-else>
-							<table class="statisticsTable schemaStats">
-								<thead>
-									<tr>
-										<th>{{ t('openregister', 'Name') }}</th>
-										<th>{{ t('openregister', 'Type') }}</th>
-										<th>{{ t('openregister', 'Actions') }}</th>
-									</tr>
-								</thead>
-								<tbody>
-									<tr v-for="(property, key) in sortedProperties(schema)" :key="key">
-										<td>{{ key }}</td>
-										<td>{{ property.type }}</td>
-										<td>
-											<NcActions :primary="false">
-												<NcActionButton close-after-click
-													:aria-label="'Edit ' + key"
-													@click="schemaStore.setSchemaPropertyKey(key); schemaStore.setSchemaItem(schema); navigationStore.setModal('editSchemaProperty')">
-													<template #icon>
-														<Pencil :size="16" />
-													</template>
-													Edit
-												</NcActionButton>
-												<NcActionButton close-after-click
-													:aria-label="'Delete ' + key"
-													@click="schemaStore.setSchemaPropertyKey(key); schemaStore.setSchemaItem(schema); navigationStore.setModal('deleteSchemaProperty')">
-													<template #icon>
-														<TrashCanOutline :size="16" />
-													</template>
-													Delete
-												</NcActionButton>
-											</NcActions>
-										</td>
-									</tr>
-									<tr v-if="!Object.keys(schema.properties).length">
-										<td colspan="3">
-											No properties found
-										</td>
-									</tr>
-								</tbody>
-							</table>
-						</div>
-
-						<!-- Toggle button -->
-						<NcButton @click="schema.showProperties = !schema.showProperties">
+					<NcActions
+						:force-name="true"
+						:inline="1"
+						:primary="true"
+						menu-name="Schema actions">
+						<NcActionButton close-after-click @click="schemaStore.setSchemaItem(null); navigationStore.setModal('editSchema')">
 							<template #icon>
-								<TableIcon v-if="schema.showProperties" :size="20" />
-								<ListIcon v-else :size="20" />
+								<PlusCircleOutline :size="20" />
 							</template>
-							{{ schema.showProperties ? 'Show Stats' : 'Show Properties' }}
-						</NcButton>
+							Add Schema
+						</NcActionButton>
+						<NcActionButton close-after-click @click="schemaStore.refreshSchemaList()">
+							<template #icon>
+								<Refresh :size="20" />
+							</template>
+							Refresh
+						</NcActionButton>
+					</NcActions>
+				</div>
+			</div>
+
+			<!-- Empty State -->
+			<NcEmptyContent v-if="!schemaStore.schemaList.length"
+				:name="t('openregister', 'No schemas found')"
+				icon="icon-folder" />
+
+			<!-- Content -->
+			<div v-else class="dashboardContent">
+				<template v-if="schemaStore.viewMode === 'cards'">
+					<div class="cardGrid">
+						<div v-for="schema in schemaStore.schemaList" :key="schema.id" class="card">
+							<div class="cardHeader">
+								<h2>
+									<FileTreeOutline :size="20" />
+									{{ schema.title }}
+								</h2>
+								<NcActions :primary="true" menu-name="Actions">
+									<template #icon>
+										<DotsHorizontal :size="20" />
+									</template>
+									<NcActionButton close-after-click @click="schemaStore.setSchemaItem(schema); navigationStore.setModal('editSchema')">
+										<template #icon>
+											<Pencil :size="20" />
+										</template>
+										Edit
+									</NcActionButton>
+									<NcActionButton close-after-click @click="schemaStore.setSchemaPropertyKey(null); schemaStore.setSchemaItem(schema); navigationStore.setModal('editSchemaProperty')">
+										<template #icon>
+											<PlusCircleOutline :size="20" />
+										</template>
+										Add Property
+									</NcActionButton>
+									<NcActionButton close-after-click @click="schemaStore.downloadSchema(schema)">
+										<template #icon>
+											<Download :size="20" />
+										</template>
+										Download
+									</NcActionButton>
+									<NcActionButton v-tooltip="schema.stats?.objects?.total > 0 ? 'Cannot delete: objects are still attached' : ''"
+										close-after-click
+										:disabled="schema.stats?.objects?.total > 0"
+										@click="schemaStore.setSchemaItem(schema); navigationStore.setDialog('deleteSchema')">
+										<template #icon>
+											<TrashCanOutline :size="20" />
+										</template>
+										Delete
+									</NcActionButton>
+									<NcActionButton close-after-click @click="schemaStore.setSchemaItem(schema); navigationStore.setSelected('schemaDetails')">
+										<template #icon>
+											<InformationOutline :size="20" />
+										</template>
+										View Details
+									</NcActionButton>
+								</NcActions>
+							</div>
+							<!-- Toggle between stats and properties -->
+							<div v-if="!schema.showProperties">
+								<table class="statisticsTable schemaStats">
+									<thead>
+										<tr>
+											<th>{{ t('openregister', 'Type') }}</th>
+											<th>{{ t('openregister', 'Total') }}</th>
+											<th>{{ t('openregister', 'Size') }}</th>
+										</tr>
+									</thead>
+									<tbody>
+										<tr>
+											<td>{{ t('openregister', 'Registers') }}</td>
+											<td>{{ schema.stats?.registers ?? 0 }}</td>
+											<td>-</td>
+										</tr>
+										<tr>
+											<td>{{ t('openregister', 'Properties') }}</td>
+											<td>{{ Object.keys(schema.properties).length }}</td>
+											<td>-</td>
+										</tr>
+										<tr>
+											<td>{{ t('openregister', 'Objects') }}</td>
+											<td>{{ schema.stats?.objects?.total || 0 }}</td>
+											<td>{{ formatBytes(schema.stats?.objects?.size || 0) }}</td>
+										</tr>
+										<tr class="subRow">
+											<td class="indented">
+												{{ t('openregister', 'Invalid') }}
+											</td>
+											<td>{{ schema.stats?.objects?.invalid || 0 }}</td>
+											<td>-</td>
+										</tr>
+										<tr class="subRow">
+											<td class="indented">
+												{{ t('openregister', 'Deleted') }}
+											</td>
+											<td>{{ schema.stats?.objects?.deleted || 0 }}</td>
+											<td>-</td>
+										</tr>
+										<tr class="subRow">
+											<td class="indented">
+												{{ t('openregister', 'Locked') }}
+											</td>
+											<td>{{ schema.stats?.objects?.locked || 0 }}</td>
+											<td>-</td>
+										</tr>
+										<tr class="subRow">
+											<td class="indented">
+												{{ t('openregister', 'Published') }}
+											</td>
+											<td>{{ schema.stats?.objects?.published || 0 }}</td>
+											<td>-</td>
+										</tr>
+										<tr>
+											<td>{{ t('openregister', 'Logs') }}</td>
+											<td>{{ schema.stats?.logs?.total || 0 }}</td>
+											<td>{{ formatBytes(schema.stats?.logs?.size || 0) }}</td>
+										</tr>
+										<tr>
+											<td>{{ t('openregister', 'Files') }}</td>
+											<td>{{ schema.stats?.files?.total || 0 }}</td>
+											<td>{{ formatBytes(schema.stats?.files?.size || 0) }}</td>
+										</tr>
+									</tbody>
+								</table>
+							</div>
+							<div v-else>
+								<table class="statisticsTable schemaStats">
+									<thead>
+										<tr>
+											<th>{{ t('openregister', 'Name') }}</th>
+											<th>{{ t('openregister', 'Type') }}</th>
+											<th>{{ t('openregister', 'Actions') }}</th>
+										</tr>
+									</thead>
+									<tbody>
+										<tr v-for="(property, key) in sortedProperties(schema)" :key="key">
+											<td>{{ key }}</td>
+											<td>{{ property.type }}</td>
+											<td>
+												<NcActions :primary="false">
+													<NcActionButton close-after-click
+														:aria-label="'Edit ' + key"
+														@click="schemaStore.setSchemaPropertyKey(key); schemaStore.setSchemaItem(schema); navigationStore.setModal('editSchemaProperty')">
+														<template #icon>
+															<Pencil :size="16" />
+														</template>
+														Edit
+													</NcActionButton>
+													<NcActionButton close-after-click
+														:aria-label="'Delete ' + key"
+														@click="schemaStore.setSchemaPropertyKey(key); schemaStore.setSchemaItem(schema); navigationStore.setModal('deleteSchemaProperty')">
+														<template #icon>
+															<TrashCanOutline :size="16" />
+														</template>
+														Delete
+													</NcActionButton>
+												</NcActions>
+											</td>
+										</tr>
+										<tr v-if="!Object.keys(schema.properties).length">
+											<td colspan="3">
+												No properties found
+											</td>
+										</tr>
+									</tbody>
+								</table>
+							</div>
+
+							<!-- Toggle button -->
+							<NcButton @click="schema.showProperties = !schema.showProperties">
+								<template #icon>
+									<TableIcon v-if="schema.showProperties" :size="20" />
+									<ListIcon v-else :size="20" />
+								</template>
+								{{ schema.showProperties ? 'Show Stats' : 'Show Properties' }}
+							</NcButton>
+						</div>
 					</div>
 				</template>
 				<template v-else>
-					<table class="statisticsTable registerStats tableOfContents">
-						<thead>
-							<tr>
-								<th>Title</th>
-								<th>Objects (Total/Size)</th>
-								<th>Logs (Total/Size)</th>
-								<th>Files (Total/Size)</th>
-								<th>Registers</th>
-								<th>Created</th>
-								<th>Updated</th>
-								<th />
-							</tr>
-						</thead>
-						<tbody>
-							<tr v-for="schema in schemaStore.schemaList" :key="schema.id">
-								<td>{{ schema.title }}</td>
-								<td>{{ schema.stats?.objects?.total || 0 }}/{{ formatBytes(schema.stats?.objects?.size || 0) }}</td>
-								<td>{{ schema.stats?.logs?.total || 0 }}/{{ formatBytes(schema.stats?.logs?.size || 0) }}</td>
-								<td>{{ schema.stats?.files?.total || 0 }}/{{ formatBytes(schema.stats?.files?.size || 0) }}</td>
-								<td>{{ schema.stats?.registers|| 0 }}</td>
-								<td>{{ schema.created ? new Date(schema.created).toLocaleDateString({day: '2-digit', month: '2-digit', year: 'numeric'}) + ', ' + new Date(schema.created).toLocaleTimeString({hour: '2-digit', minute: '2-digit', second: '2-digit'}) : '-' }}</td>
-								<td>{{ schema.updated ? new Date(schema.updated).toLocaleDateString({day: '2-digit', month: '2-digit', year: 'numeric'}) + ', ' + new Date(schema.updated).toLocaleTimeString({hour: '2-digit', minute: '2-digit', second: '2-digit'}) : '-' }}</td>
-								<td>
-									<NcActions :primary="false">
-										<template #icon>
-											<DotsHorizontal :size="20" />
-										</template>
-										<NcActionButton close-after-click @click="schemaStore.setSchemaItem(schema); navigationStore.setModal('editSchema')">
+					<div class="viewTableContainer">
+						<table class="viewTable registerStats tableOfContents">
+							<thead>
+								<tr>
+									<th>Title</th>
+									<th>Objects (Total/Size)</th>
+									<th>Logs (Total/Size)</th>
+									<th>Files (Total/Size)</th>
+									<th>Registers</th>
+									<th>Created</th>
+									<th>Updated</th>
+									<th class="tableColumnActions">
+										{{ t('openregister', 'Actions') }}
+									</th>
+								</tr>
+							</thead>
+							<tbody>
+								<tr v-for="schema in schemaStore.schemaList" :key="schema.id" class="viewTableRow">
+									<td class="tableColumnTitle">
+										<div class="titleContent">
+											<strong>{{ schema.title }}</strong>
+											<span v-if="schema.description" class="textDescription textEllipsis">{{ schema.description }}</span>
+										</div>
+									</td>
+									<td>{{ schema.stats?.objects?.total || 0 }}/{{ formatBytes(schema.stats?.objects?.size || 0) }}</td>
+									<td>{{ schema.stats?.logs?.total || 0 }}/{{ formatBytes(schema.stats?.logs?.size || 0) }}</td>
+									<td>{{ schema.stats?.files?.total || 0 }}/{{ formatBytes(schema.stats?.files?.size || 0) }}</td>
+									<td>{{ schema.stats?.registers|| 0 }}</td>
+									<td>{{ schema.created ? new Date(schema.created).toLocaleDateString({day: '2-digit', month: '2-digit', year: 'numeric'}) + ', ' + new Date(schema.created).toLocaleTimeString({hour: '2-digit', minute: '2-digit', second: '2-digit'}) : '-' }}</td>
+									<td>{{ schema.updated ? new Date(schema.updated).toLocaleDateString({day: '2-digit', month: '2-digit', year: 'numeric'}) + ', ' + new Date(schema.updated).toLocaleTimeString({hour: '2-digit', minute: '2-digit', second: '2-digit'}) : '-' }}</td>
+									<td class="tableColumnActions">
+										<NcActions :primary="false">
 											<template #icon>
-												<Pencil :size="20" />
+												<DotsHorizontal :size="20" />
 											</template>
-											Edit
-										</NcActionButton>
-										<NcActionButton close-after-click @click="schemaStore.setSchemaPropertyKey(null); schemaStore.setSchemaItem(schema); navigationStore.setModal('editSchemaProperty')">
-											<template #icon>
-												<PlusCircleOutline :size="20" />
-											</template>
-											Add Property
-										</NcActionButton>
-										<NcActionButton close-after-click @click="schemaStore.downloadSchema(schema)">
-											<template #icon>
-												<Download :size="20" />
-											</template>
-											Download
-										</NcActionButton>
-										<NcActionButton v-tooltip="schema.stats?.objects?.total > 0 ? 'Cannot delete: objects are still attached' : ''"
-											close-after-click
-											:disabled="schema.stats?.objects?.total > 0"
-											@click="schemaStore.setSchemaItem(schema); navigationStore.setDialog('deleteSchema')">
-											<template #icon>
-												<TrashCanOutline :size="20" />
-											</template>
-											Delete
-										</NcActionButton>
-										<NcActionButton close-after-click @click="schemaStore.setSchemaItem(schema); navigationStore.setSelected('schemaDetails')">
-											<template #icon>
-												<InformationOutline :size="20" />
-											</template>
-											View Details
-										</NcActionButton>
-									</NcActions>
-								</td>
-							</tr>
-						</tbody>
-					</table>
+											<NcActionButton close-after-click @click="schemaStore.setSchemaItem(schema); navigationStore.setModal('editSchema')">
+												<template #icon>
+													<Pencil :size="20" />
+												</template>
+												Edit
+											</NcActionButton>
+											<NcActionButton close-after-click @click="schemaStore.setSchemaPropertyKey(null); schemaStore.setSchemaItem(schema); navigationStore.setModal('editSchemaProperty')">
+												<template #icon>
+													<PlusCircleOutline :size="20" />
+												</template>
+												Add Property
+											</NcActionButton>
+											<NcActionButton close-after-click @click="schemaStore.downloadSchema(schema)">
+												<template #icon>
+													<Download :size="20" />
+												</template>
+												Download
+											</NcActionButton>
+											<NcActionButton v-tooltip="schema.stats?.objects?.total > 0 ? 'Cannot delete: objects are still attached' : ''"
+												close-after-click
+												:disabled="schema.stats?.objects?.total > 0"
+												@click="schemaStore.setSchemaItem(schema); navigationStore.setDialog('deleteSchema')">
+												<template #icon>
+													<TrashCanOutline :size="20" />
+												</template>
+												Delete
+											</NcActionButton>
+											<NcActionButton close-after-click @click="schemaStore.setSchemaItem(schema); navigationStore.setSelected('schemaDetails')">
+												<template #icon>
+													<InformationOutline :size="20" />
+												</template>
+												View Details
+											</NcActionButton>
+										</NcActions>
+									</td>
+								</tr>
+							</tbody>
+						</table>
+					</div>
 				</template>
 			</div>
 		</div>
@@ -358,17 +382,51 @@ export default {
 }
 </script>
 <style scoped lang="scss">
-.pageHeaderContainer {
+.viewContainer {
 	display: flex;
-	align-items: center;
-	justify-content: space-between;
-	padding: 0;
+	flex-direction: column;
+	height: 100%;
 }
 
-.pageHeader {
+.viewHeader {
+	padding: 20px;
+	background-color: var(--color-primary-light);
+	color: var(--color-main-text);
+}
+
+.viewHeaderTitleIndented {
 	font-size: 30px;
 	font-weight: 600;
 	margin-left: 50px;
+}
+
+.viewActionsBar {
+	display: flex;
+	align-items: end;
+	gap: 1em;
+	padding-block-start: 0.5em;
+	padding-inline-end: 0.9em;
+}
+
+.viewInfo {
+	display: flex;
+	align-items: center;
+}
+
+.viewTotalCount {
+	font-size: 1.2em;
+	font-weight: 600;
+}
+
+.viewActions {
+	display: flex;
+	align-items: end;
+	gap: 1em;
+}
+
+.viewModeSwitchContainer {
+	display: flex;
+	align-items: center;
 }
 
 .dashboardContent {
@@ -378,13 +436,13 @@ export default {
 	padding-inline: 20px;
 }
 
-.registers {
+.cardGrid {
 	display: grid;
 	gap: 1.5rem;
 	grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
 }
 
-.registerCard {
+.card {
 	background: var(--color-main-background);
 	border-radius: 8px;
 	padding-block: 20px;
@@ -395,7 +453,7 @@ export default {
 	border: 1px solid var(--color-border);
 }
 
-.registerHeader {
+.cardHeader {
 	display: flex;
 	align-items: center;
 	justify-content: space-between;
@@ -490,22 +548,7 @@ export default {
 .statisticsTable.schemaStats tr:last-child td {
 	border-block-end: none;
 }
-.headerActionsContainer {
-    display: flex;
-    align-items: end;
-    gap: 1em;
-    padding-block-start: 0.5em;
-    padding-inline-end: 0.9em;
 
-    .viewModeSwitchContainer {
-        display: flex;
-        align-items: center;
-
-        span {
-            max-height: 34px;
-        }
-    }
-}
 .tableOfContents {
 	thead {
 		th {
@@ -521,5 +564,52 @@ export default {
 			background-color: var(--color-border);
 		}
 	}
+}
+
+.viewTableContainer {
+	padding: 20px;
+}
+
+.viewTable {
+	width: 100%;
+	border-collapse: collapse;
+	font-size: 0.9em;
+	background: var(--color-main-background);
+	border-radius: var(--border-radius);
+	overflow: hidden;
+}
+
+.viewTableRow {
+	transition: background-color 0.2s ease-in-out;
+}
+
+.viewTableRow:hover {
+	background-color: var(--color-background-hover);
+}
+
+.tableColumnTitle {
+	display: flex;
+	align-items: center;
+	gap: 8px;
+}
+
+.titleContent {
+	display: flex;
+	flex-direction: column;
+}
+
+.textDescription {
+	color: var(--color-text-maxcontrast);
+	font-size: 0.9em;
+}
+
+.textEllipsis {
+	white-space: nowrap;
+	overflow: hidden;
+	text-overflow: ellipsis;
+}
+
+.tableColumnActions {
+	text-align: right;
 }
 </style>
